+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @workspace/hooks
 
+## 0.0.2
+
+### Patch Changes
+
+- New hooks use-bluetooth et use-set
+
 ## 0.0.1
 
 ### Patch Changes